:alien: fix: New API requires authentication
The new updated API now requires authentication for all endpoints.
The authentication use the "Authorization" header with a JWT token in
the format below:
```
Authorization: Bearer ${TOKEN}
```
The token is obtained, by signing in using AWS Amplify.

Also note that we're using an axios instance with a `baseUrl` in a
dedicated module, to bypass CORS issues.